Move core agent modules into scout_apm.core.agent
Keep them all together in one place
