+++ src/scout_apm/core/agent/commands.py
