+++ src/scout_apm/core/agent/manager.py
@@ -1,5 +1,4 @@
 # coding=utf-8
-# coding=utf-8
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import errno
